split property_expressions.go
