+++ property_json_expression.go
@@ -1,0 +1,83 @@
+package qradar
+
+import (
+	"context"
+	"net/http"
+)
+
+// PropertyJSONExpressionService handles methods related to Property JSON Expressions of the QRadar API.
+type PropertyJSONExpressionService service
+
+const (
+	propertyJSONExpressionAPIPrefix = "api/config/event_sources/custom_properties/property_json_expressions"
+)
+
+// Get returns Property JSON Expressions of the current QRadar installation
+func (c *PropertyJSONExpressionService) Get(ctx context.Context, fields, filter string, from, to int) ([]PropertyExpression, error) {
+	req, err := c.client.requestHelp(http.MethodGet, propertyJSONExpressionAPIPrefix, fields, filter, from, to, nil, nil)
+	if err != nil {
+		return nil, err
+	}
+	var result []PropertyExpression
+	_, err = c.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, err
+	}
+	return result, nil
+}
+
+// GetByID returns Property JSON Expression of the current QRadar installation by ID.
+func (c *PropertyJSONExpressionService) GetByID(ctx context.Context, fields string, id int) (*PropertyExpression, error) {
+	req, err := c.client.requestHelp(http.MethodGet, propertyJSONExpressionAPIPrefix, fields, "", 0, 0, &id, nil)
+	if err != nil {
+		return nil, err
+	}
+	var result PropertyExpression
+	_, err = c.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
+
+// Create creates Property JSON Expression in QRadar installation>
+func (c *PropertyJSONExpressionService) Create(ctx context.Context, fields string, data interface{}) (*PropertyExpression, error) {
+	req, err := c.client.requestHelp(http.MethodPost, propertyJSONExpressionAPIPrefix, fields, "", 0, 0, nil, data)
+	if err != nil {
+		return nil, err
+	}
+	var result PropertyExpression
+	_, err = c.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
+
+// UpdateByID updates Property JSON Expression in QRadar installation by ID.
+func (c *PropertyJSONExpressionService) UpdateByID(ctx context.Context, fields string, id int, data interface{}) (*PropertyExpression, error) {
+	req, err := c.client.requestHelp(http.MethodPost, propertyJSONExpressionAPIPrefix, fields, "", 0, 0, &id, data)
+	if err != nil {
+		return nil, err
+	}
+	var result PropertyExpression
+	_, err = c.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
+
+// DeleteByID creates A Delete Task in QRadar installation in order to safely delete Property JSON Expression by ID.
+func (c *PropertyJSONExpressionService) DeleteByID(ctx context.Context, fields string, id int) (*DeleteTask, error) {
+	req, err := c.client.requestHelp(http.MethodDelete, propertyJSONExpressionAPIPrefix, fields, "", 0, 0, &id, nil)
+	if err != nil {
+		return nil, err
+	}
+	var result DeleteTask
+	_, err = c.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result, nil
+}
